+++ lib/src/widgets/custom_image_crop_widget.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:math';
 import 'dart:ui' as ui;
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:gesture_x_detector/gesture_x_detector.dart';
 import 'package:vector_math/vector_math_64.dart' as vector_math;
@@ -12,17 +11,57 @@ import 'package:custom_image_crop/src/painters/dotted_path_painter.dart';
 import 'package:custom_image_crop/src/clippers/inverted_clipper.dart';
 import 'package:custom_image_crop/src/models/model.dart';
 
+/// An image cropper that is customizable.
+/// You can rotate, scale and translate either
+/// through gestures or a controller
 class CustomImageCrop extends StatefulWidget {
+  /// The image to crop
   final ImageProvider image;
+
+  /// The controller that handles the cropping and
+  /// changing of the cropping area
   final CustomImageCropController cropController;
+
+  /// The color behind the cropping area
   final Color backgroundColor;
+
+  /// The color in front of the cropped area
   final Color overlayColor;
+
+  /// The shape of the cropping area
   final CustomCropShape shape;
   final CustomCropShape outputShape;
+
+  /// The percentage of the available area that is
+  /// reserved for the cropping area
   final double cropPercentage;
+
+  /// The path drawer of the border see [DottedCropPathPainter],
+  /// [SolidPathPainter] for more details or how to implement a
+  /// custom one
   final CustomPaint Function(Path) drawPath;
-  final bool canRotate;
   final double minScale;
+
+  /// Whether to allow the image to be rotated.
+  final bool canRotate;
+
+  /// Determines whether scaling gesture is disabled.
+  ///
+  /// By default, scaling is enabled.
+  /// Set [canScale] to `false` to disable scaling.
+  final bool canScale;
+
+  /// Determines whether moving gesture overlay is disabled.
+  ///
+  /// By default, moving is enabled.
+  /// Set [canMove] to `false` to disable move.
+  final bool canMove;
+
+  /// The paint used when drawing an image before cropping
+  final Paint imagePaintDuringCrop;
+
+  /// This widget is used to specify a custom progress indicator
+  final Widget? customProgressIndicator;
 
   /// A custom image cropper widget
   ///
@@ -40,19 +79,25 @@ class CustomImageCrop extends StatefulWidget {
   /// we've provided two default painters as inspiration
   /// `DottedCropPathPainter.drawPath` and
   /// `SolidCropPathPainter.drawPath`
-  const CustomImageCrop({
+  CustomImageCrop({
     required this.image,
     required this.cropController,
     this.overlayColor = const Color.fromRGBO(0, 0, 0, 0.5),
     this.backgroundColor = Colors.white,
-    this.shape = CustomCropShape.circle,
-    this.outputShape = CustomCropShape.square,
+    this.shape = CustomCropShape.Circle,
+    this.outputShape = CustomCropShape.Square,
     this.cropPercentage = 0.8,
     this.drawPath = DottedCropPathPainter.drawPath,
     this.canRotate = true,
+    this.canScale = true,
+    this.canMove = true,
+    this.customProgressIndicator,
+    Paint? imagePaintDuringCrop,
     this.minScale = 1.0,
     Key? key,
-  }) : super(key: key);
+  })  : this.imagePaintDuringCrop = imagePaintDuringCrop ??
+            (Paint()..filterQuality = FilterQuality.high),
+        super(key: key);
 
   @override
   _CustomImageCropState createState() => _CustomImageCropState();
@@ -60,10 +105,10 @@ class CustomImageCrop extends StatefulWidget {
 
 class _CustomImageCropState extends State<CustomImageCrop>
     with CustomImageCropListener {
-  CropImageData? dataTransitionStart;
-  late Path path;
-  late double width, height;
-  ui.Image? imageAsUIImage;
+  CropImageData? _dataTransitionStart;
+  late Path _path;
+  late double _width, _height;
+  ui.Image? _imageAsUIImage;
   ImageStream? _imageStream;
   ImageStreamListener? _imageListener;
 
@@ -93,7 +138,7 @@ class _CustomImageCropState extends State<CustomImageCrop>
 
   void _updateImage(ImageInfo imageInfo, _) {
     setState(() {
-      imageAsUIImage = imageInfo.image;
+      _imageAsUIImage = imageInfo.image;
     });
   }
 
@@ -108,40 +153,39 @@ class _CustomImageCropState extends State<CustomImageCrop>
 
   @override
   Widget build(BuildContext context) {
-    final image = imageAsUIImage;
+    final image = _imageAsUIImage;
     if (image == null) {
-      return const Center(child: CircularProgressIndicator());
+      return Center(
+        child: widget.customProgressIndicator ?? CircularProgressIndicator(),
+      );
     }
     return LayoutBuilder(
       builder: (context, constraints) {
-        width = constraints.maxWidth;
-        height = constraints.maxHeight;
-        final cropWidth = min(width, height) * widget.cropPercentage;
-        final imageWidth = min(image.width, image.height);
-        // final defaultScale = min(image.width, image.height) / cropWidth;
-        // final scale = data.scale * defaultScale;
-        final defaultScale = cropWidth / imageWidth;
+        _width = constraints.maxWidth;
+        _height = constraints.maxHeight;
+        final cropWidth = min(_width, _height) * widget.cropPercentage;
+        final defaultScale = cropWidth / max(image.width, image.height);
         final scale = data.scale * defaultScale;
-        path = _getPath(cropWidth, width, height);
+        _path = _getPath(cropWidth, _width, _height);
         return XGestureDetector(
           onMoveStart: onMoveStart,
           onMoveUpdate: onMoveUpdate,
           onScaleStart: onScaleStart,
           onScaleUpdate: onScaleUpdate,
           child: Container(
-            width: width,
-            height: height,
+            width: _width,
+            height: _height,
             color: widget.backgroundColor,
             child: Stack(
               children: [
                 Positioned(
-                  left: data.x + width / 2,
-                  top: data.y + height / 2,
+                  left: data.x + _width / 2,
+                  top: data.y + _height / 2,
                   child: Transform(
-                    transform:
-                        Matrix4.diagonal3(vector_math.Vector3(scale, scale, 0))
-                          ..rotateZ(data.angle)
-                          ..translate(-image.width / 2, -image.height / 2),
+                    transform: Matrix4.diagonal3(
+                        vector_math.Vector3(scale, scale, scale))
+                      ..rotateZ(data.angle)
+                      ..translate(-image.width / 2, -image.height / 2),
                     child: Image(
                       image: widget.image,
                     ),
@@ -149,13 +193,13 @@ class _CustomImageCropState extends State<CustomImageCrop>
                 ),
                 IgnorePointer(
                   child: ClipPath(
-                    clipper: InvertedClipper(path, width, height),
+                    clipper: InvertedClipper(_path, _width, _height),
                     child: Container(
                       color: widget.overlayColor,
                     ),
                   ),
                 ),
-                widget.drawPath(path),
+                widget.drawPath(_path),
               ],
             ),
           ),
@@ -165,39 +209,43 @@ class _CustomImageCropState extends State<CustomImageCrop>
   }
 
   void onScaleStart(_) {
-    dataTransitionStart = null; // Reset for update
+    _dataTransitionStart = null; // Reset for update
   }
 
   void onScaleUpdate(ScaleEvent event) {
-    if (dataTransitionStart != null) {
-      addTransition(dataTransitionStart! -
-          CropImageData(
-            scale: event.scale,
-            angle: widget.canRotate ? event.rotationAngle : 0,
-          ));
+    final scale =
+        widget.canScale ? event.scale : (_dataTransitionStart?.scale ?? 1.0);
+
+    final angle = widget.canRotate ? event.rotationAngle : 0.0;
+
+    if (_dataTransitionStart != null) {
+      addTransition(
+        _dataTransitionStart! -
+            CropImageData(
+              scale: scale,
+              angle: angle,
+            ),
+      );
     }
-    dataTransitionStart = CropImageData(
-      scale: event.scale,
-      angle: widget.canRotate ? event.rotationAngle : 0,
+    _dataTransitionStart = CropImageData(
+      scale: scale,
+      angle: angle,
     );
   }
 
   void onMoveStart(_) {
-    dataTransitionStart = null; // Reset for update
+    _dataTransitionStart = null; // Reset for update
   }
 
   void onMoveUpdate(MoveEvent event) {
+    if (!widget.canMove) return;
+
     addTransition(CropImageData(x: event.delta.dx, y: event.delta.dy));
   }
 
-  Path _getPath(
-    double cropWidth,
-    double width,
-    double height, {
-    CustomCropShape? shape,
-  }) {
-    switch (shape ?? widget.shape) {
-      case CustomCropShape.circle:
+  Path _getPath(double cropWidth, double width, double height) {
+    switch (widget.shape) {
+      case CustomCropShape.Circle:
         return Path()
           ..addOval(
             Rect.fromCircle(
@@ -219,29 +267,23 @@ class _CustomImageCropState extends State<CustomImageCrop>
 
   @override
   Future<MemoryImage?> cropImage() async {
-    final image = imageAsUIImage;
-    if (image == null) {
+    if (_imageAsUIImage == null) {
       return null;
     }
-    final cropWidth = min(width, height) * widget.cropPercentage;
+    final imageWidth = _imageAsUIImage!.width;
+    final imageHeight = _imageAsUIImage!.height;
     final pictureRecorder = ui.PictureRecorder();
     final canvas = Canvas(pictureRecorder);
-    // final defaultScale = min(image.width, image.height) / cropWidth;
-    // final scale = data.scale * defaultScale;
-    final imageWidth = min(image.width, image.height);
-    final defaultScale = cropWidth / imageWidth;
-    final scale = data.scale * defaultScale;
-    final clipPath = Path.from(_getPath(
-      cropWidth,
-      cropWidth,
-      cropWidth,
-      shape: widget.outputShape,
-    ));
-    final matrix4Image = Matrix4.diagonal3(vector_math.Vector3(1, 1, 0))
-      ..translate(data.x + cropWidth / 2, data.y + cropWidth / 2)
+    final uiWidth = min(_width, _height) * widget.cropPercentage;
+    final cropWidth = max(imageWidth, imageHeight).toDouble();
+    final translateScale = cropWidth / uiWidth;
+    final scale = data.scale;
+    final clipPath = Path.from(_getPath(cropWidth, cropWidth, cropWidth));
+    final matrix4Image = Matrix4.diagonal3(vector_math.Vector3.all(1))
+      ..translate(translateScale * data.x + cropWidth / 2,
+          translateScale * data.y + cropWidth / 2)
       ..scale(scale)
       ..rotateZ(data.angle);
-    final imagePaint = Paint()..isAntiAlias = false;
     final bgPaint = Paint()
       ..color = widget.backgroundColor
       ..style = PaintingStyle.fill;
@@ -249,11 +291,8 @@ class _CustomImageCropState extends State<CustomImageCrop>
     canvas.save();
     canvas.clipPath(clipPath);
     canvas.transform(matrix4Image.storage);
-    canvas.drawImage(
-      image,
-      Offset(-image.width / 2, -image.height / 2),
-      imagePaint,
-    );
+    canvas.drawImage(_imageAsUIImage!,
+        Offset(-imageWidth / 2, -imageHeight / 2), widget.imagePaintDuringCrop);
     canvas.restore();
 
     // Optionally remove magenta from image by evaluating every pixel
@@ -262,12 +301,12 @@ class _CustomImageCropState extends State<CustomImageCrop>
     // final bytes = await compute(computeToByteData, <String, dynamic>{'pictureRecorder': pictureRecorder, 'cropWidth': cropWidth});
 
     ui.Picture picture = pictureRecorder.endRecording();
-    ui.Image outputImage =
+    ui.Image image =
         await picture.toImage(cropWidth.floor(), cropWidth.floor());
+
     // Adding compute would be preferable. Unfortunately we cannot pass an ui image to this.
     // A workaround would be to save the image and load it inside of the isolate
-    final bytes = await outputImage.toByteData(format: ui.ImageByteFormat.png);
-
+    final bytes = await image.toByteData(format: ui.ImageByteFormat.png);
     return bytes == null ? null : MemoryImage(bytes.buffer.asUint8List());
   }
 
@@ -295,6 +334,6 @@ class _CustomImageCropState extends State<CustomImageCrop>
 }
 
 enum CustomCropShape {
-  circle,
-  square,
+  Circle,
+  Square,
 }