+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.0.4] - 25/01/2021
+
+* Added documentation
+
+## [0.0.3] - 13/12/2021
+
+* Fixed issue where cropped image's size depends on screen size used
+* Fixed issue where cropped image's quality is worse than original image
+* Updated to flutter 2.8.0
+
 ## [0.0.2] - 30/03/2021
 
 * Updated docs